+++ ocs_ci/framework/pytest_customization/ocscilib.py
@@ -18,12 +18,11 @@ from ocs_ci.utility.utils import (
     dump_config_to_file,
     get_cluster_version,
     get_ceph_version,
-    get_rook_version,
     get_csi_versions,
     get_testrun_name,
 )
 from ocs_ci.ocs.utils import collect_ocs_logs
-from ocs_ci.ocs.resources.pod import get_version_info
+from ocs_ci.ocs.resources.ocs import get_version_info
 from ocs_ci.ocs.resources.catalog_source import CatalogSource
 from ocs_ci.ocs.constants import (
     OPERATOR_CATALOG_SOURCE_NAME,
@@ -167,9 +166,7 @@ def pytest_configure(config):
             )
             skip_list = ['ocs-operator']
             for key, val in mods.items():
-                if key in skip_list:
-                    pass
-                else:
+                if key not in skip_list:
                     config._metadata[key] = val.rsplit('/')[-1]
         except (FileNotFoundError, CommandFailed):
             pass
